// app.client/src/api/models/InventoryItemListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InventoryItemListVM = {
    inventoryItemId?: string | null;
    ownerBarcodeItemNum?: string | null;
    productClientId?: string | null;
    productDescription?: string | null;
    productNotes?: string | null;
    productUom?: string | null;
    productGrossWeightKg?: number;
    productPackage?: string | null;
    lastPricePaid?: number;
    isLotRequired?: boolean;
    isSnRequired?: boolean;
    isDateMfgRequired?: boolean;
    isDateExpRequired?: boolean;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    imageUrl?: string | null;
    isInTransaction?: boolean;
    inventoryItemTypeId?: string | null;
    inventoryItemTypeName?: string | null;
};